# ARROW-764: [C++] Improves performance of CopyBitmap and adds benchmarks