--- a/cpp/src/arrow/util/CMakeLists.txt
+++ b/cpp/src/arrow/util/CMakeLists.txt
@@ -63,10 +63,10 @@ if (ARROW_BUILD_BENCHMARKS)
       Shlwapi.lib
   )
   else()
-	  target_link_libraries(arrow_benchmark_main
+    target_link_libraries(arrow_benchmark_main
       benchmark
       pthread
-	  )
+    )
   endif()
 
   # TODO(wesm): Some benchmarks include gtest.h
@@ -79,5 +79,7 @@ ADD_ARROW_TEST(decimal-test)
 ADD_ARROW_TEST(key-value-metadata-test)
 ADD_ARROW_TEST(rle-encoding-test)
 ADD_ARROW_TEST(stl-util-test)
+
+ADD_ARROW_BENCHMARK(bit-util-benchmark)
 
 add_subdirectory(variant)