@seibs 2017-12-22
Thanks for taking a look @xhochy.

My latest push fixes the formatting issues and also adds a test+fix for a few corner cases my first upload didn't handle correctly.  

**Benchmark with recent changes**
```
Run on (4 X 2208 MHz CPU s)
12/22/17 15:14:31
Benchmark                                        Time           CPU Iterations
------------------------------------------------------------------------------
BM_CopyBitmap/97.6563k/0/min_time:1.000          5 us          5 us     270310   17.9019GB/s
BM_CopyBitmap/976.563k/0/min_time:1.000         65 us         65 us      21854   14.3143GB/s
BM_CopyBitmap/97.6563k/4/min_time:1.000        153 us        152 us       8960   628.592MB/s
BM_CopyBitmap/976.563k/4/min_time:1.000       1534 us       1535 us        896   621.449MB/s
```

w.r.t. the comments above in `CMakeLists.txt`, I actually have a separate pull request open for them at #1406.