@xhochy 2017-12-30
@seibs Can resolve the two comments in the other pull request? I would then merge both :)