@wesm 2018-01-05
Rebased again to see if build passes...