@siddharthteotia 2018-01-05
What are the general performance concerns we are addressing here? I am just trying to understand if there is something similar to be done on JAVA side for bitvector.