@wesm 2018-01-07
@siddharthteotia there's a couple places in the codebase where we copy bits out of a bitmap, e.g. after slicing vectors https://github.com/apache/arrow/blob/2e3832f34a8933c1ef81331e6d1dde5190366852/cpp/src/arrow/ipc/writer.cc#L63. Depending on the slice offset (whether it's a multiple of 8) we have to either copy 1 bit at a time or do a memcpy